Chatbot_Information_Systems_Web/backend/requirements.txt
fastapi==0.115.2
langchain==0.3.3
langchain_community==0.3.2
langchain_core==0.3.12
langchain_groq==0.2.0
langchain_openai==0.2.2
langchain_qdrant==0.1.4
langgraph==0.2.38
passlib==1.7.4
pydantic==2.9.2
pydantic_settings==2.6.0
python_jose==3.3.0
qdrant_client==1.12.0
SQLAlchemy==2.0.35
sqlmodel==0.0.22
typing_extensions==4.12.2

sentence-transformers
asyncpg
python-multipart